+++ CHANGELOG.md
@@ -11,6 +11,36 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+- Fixed a bug with creating the correct runtime with auth using http bearer security scheme or when root auth is set. [#6357](https://github.com/microsoft/kiota/issues/6357)   
+
+## [1.24.3] - 2025-03-26
+
+### Changed
+
+- Fixed a bug with reading adaptive card OpenAPI extension and writing to plugin manifest. [#6351](https://github.com/microsoft/kiota/issues/6351)
+- Fixed a critical bug with the vscode integration and the JSON RPC server.
+
+## [1.24.2] - 2025-03-25
+
+### Added
+
+### Changed
+
+- Fixed a bug following OpenAPI 3.1 support that lead to false warnings about unknown type/format pairs. [#6270](https://github.com/microsoft/kiota/issues/6270)
+- Fixed a bug where security scheme would be missing from plugins OpenAPI descriptions. [#6279](https://github.com/microsoft/kiota/issues/6279)
+- Fixed pet store description source.
+
+## [1.24.1] - 2025-03-13
+
+### Changed
+
+- Fixed docker images publishing
+
+## [1.24.0] - 2025-03-12
+
+### Added
+
 - Added support for OpenAPI 3.1. [#3914](https://github.com/microsoft/kiota/issues/3914)
 
 ### Changed
@@ -18,6 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed mapping of Binary Types to array buffer in TypeScript. [#6124](https://github.com/microsoft/kiota/issues/6124)
 - Updated kiota generate classname flag validator to prevent invalid class names. [#6187](https://github.com/microsoft/kiota/issues/6187)
 - Fixed enum serialization for union and intersection types in Dart. [#6201](https://github.com/microsoft/kiota/pull/6201)
+- Added mapping for Bundle package in PHP to show up in Kiota commands. [#4636](https://github.com/microsoft/kiota/issues/4636)
+- Fixes javadoc warning present in generated Java code comments due to the presence of special XML character.
 
 ## [1.23.0] - 2025-02-06
 
@@ -1574,3 +1606,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Initial GitHub release
+